refactor(structure): remove directives.ts and module.ts
remove directives.ts and module.ts
